+++ app/lib/persistence/TodoCategory.scala
@@ -1,42 +1,34 @@
-/**
-  * This is a sample of Todo Application.
-  * 
-  */
 
 package lib.persistence
 
 import scala.concurrent.Future
 import ixias.persistence.SlickRepository
-import lib.model.Todo
+import lib.model.TodoCategory
 import slick.jdbc.JdbcProfile
 
-// TodoRepository: TodoTableへのクエリ発行を行うRepository層の定義
-//~~~~~~~~~~~~~~~~~~~~~~
-case class TodoRepository[P <: JdbcProfile]()(implicit val driver: P)
-  extends SlickRepository[Todo.Id, Todo, P]
+
+case class TodoCategoryRepository[P <: JdbcProfile]()(implicit val driver: P)
+  extends SlickRepository[TodoCategory.Id, TodoCategory, P]
   with db.SlickResourceProvider[P] {
 
   import api._
 
-  /**
-    * Get Todo Data
-    */
   def get(id: Id): Future[Option[EntityEmbeddedId]] =
-    RunDBAction(TodoTable, "slave") { _
+    RunDBAction(TodoCategoryTable, "slave") { _
       .filter(_.id === id)
       .result.headOption
   }
 
-  def getAll(): Future[Seq[EntityEmbeddedId]] =
-    RunDBAction(TodoTable, "slave") { _
+  def getAll(): Future[Seq[EntityEmbeddedId]] = 
+    RunDBAction(TodoCategoryTable, "slave") { _
       .result
-  }
+    }
 
   /**
     * Add Todo Data
    */
   def add(entity: EntityWithNoId): Future[Id] =
-    RunDBAction(TodoTable) { slick =>
+    RunDBAction(TodoCategoryTable) { slick =>
       slick returning slick.map(_.id) += entity.v
     }
 
@@ -44,7 +36,7 @@ case class TodoRepository[P <: JdbcProfile]()(implicit val driver: P)
    * Update Todo Data
    */
   def update(entity: EntityEmbeddedId): Future[Option[EntityEmbeddedId]] =
-    RunDBAction(TodoTable) { slick =>
+    RunDBAction(TodoCategoryTable) { slick =>
       val row = slick.filter(_.id === entity.id)
       for {
         old <- row.result.headOption
@@ -59,7 +51,7 @@ case class TodoRepository[P <: JdbcProfile]()(implicit val driver: P)
    * Delete Todo Data
    */
   def remove(id: Id): Future[Option[EntityEmbeddedId]] =
-    RunDBAction(TodoTable) { slick =>
+    RunDBAction(TodoCategoryTable) { slick =>
       val row = slick.filter(_.id === id)
       for {
         old <- row.result.headOption
